feat: refactor admin pages and enhance functionality
- Updated the project name in package.json and package-lock.json to "Portfolio".
- Removed the REST_API_SUMMARY.md file as it is no longer needed.
- Introduced new admin pages for analytics, contact management, and newsletter subscriptions with improved UI components.
- Added image upload functionality with a dedicated upload client page.
- Integrated React Scan for performance monitoring in development.
- Enhanced admin navigation to include links to new features.
- Improved OG image generation with updated metadata for better SEO.
