+++ app/admin/upload/images/page.tsx
@@ -1,326 +1,5 @@
-"use client";
+import { UploadClientPage } from "./UploadClientPage"
 
-import type { PutBlobResult } from "@vercel/blob";
-import { useState, useRef, useCallback } from "react";
-import { motion, AnimatePresence } from "framer-motion";
-import { Camera, Upload, CheckCircle, AlertCircle, X } from "lucide-react";
-
-export default function MinimalAvatarUpload() {
-  const inputFileRef = useRef<HTMLInputElement>(null);
-  const [blob, setBlob] = useState<PutBlobResult | null>(null);
-  const [dragActive, setDragActive] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [preview, setPreview] = useState<string | null>(null);
-
-  // Handle drag events
-  const handleDrag = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (e.type === "dragenter" || e.type === "dragover") {
-      setDragActive(true);
-    } else if (e.type === "dragleave") {
-      setDragActive(false);
-    }
-  }, []);
-
-  // Handle drop event
-  const handleDrop = useCallback((e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setDragActive(false);
-
-    if (e.dataTransfer.files && e.dataTransfer.files[0]) {
-      const file = e.dataTransfer.files[0];
-      handleFile(file);
-    }
-  }, []);
-
-  // Process the selected file
-  const handleFile = (file: File) => {
-    setError(null);
-
-    if (!file.type.match("image.*")) {
-      setError("Please select an image file");
-      return;
-    }
-
-    if (file.size > 5 * 1024 * 1024) {
-      setError("Image size should be less than 5MB");
-      return;
-    }
-
-    const reader = new FileReader();
-    reader.onloadend = () => {
-      setPreview(reader.result as string);
-    };
-    reader.readAsDataURL(file);
-
-    if (inputFileRef.current) {
-      const dataTransfer = new DataTransfer();
-      dataTransfer.items.add(file);
-      inputFileRef.current.files = dataTransfer.files;
-    }
-  };
-
-  // Handle file input change
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (e.target.files && e.target.files[0]) {
-      handleFile(e.target.files[0]);
-    }
-  };
-
-  // Handle form submission
-  const handleSubmit = async (event: React.FormEvent) => {
-    event.preventDefault();
-
-    if (!inputFileRef.current?.files?.length) {
-      setError("Please select an image to upload");
-      return;
-    }
-
-    try {
-      setLoading(true);
-      setError(null);
-      const file = inputFileRef.current.files[0];
-
-      const response = await fetch(`/api/avatar/upload?filename=${file.name}`, {
-        method: "POST",
-        body: file,
-      });
-
-      if (!response.ok) {
-        throw new Error(`Upload failed: ${response.statusText}`);
-      }
-
-      const newBlob = (await response.json()) as PutBlobResult;
-      setBlob(newBlob);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Upload failed");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  // Reset the form
-  const handleReset = () => {
-    setBlob(null);
-    setPreview(null);
-    setError(null);
-    if (inputFileRef.current) {
-      inputFileRef.current.value = "";
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-white p-6 flex flex-col items-center justify-center">
-      <div className="w-full max-w-md">
-        <motion.div
-          initial={{ opacity: 0, y: -20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.5 }}
-          className="bg-white border border-gray-200 rounded-lg overflow-hidden"
-        >
-          <div className="p-8">
-            <div className="flex items-center justify-center mb-8">
-              <motion.div
-                initial={{ scale: 0.8 }}
-                animate={{ scale: 1 }}
-                transition={{
-                  type: "spring",
-                  stiffness: 260,
-                  damping: 20,
-                }}
-              >
-                <Camera className="h-8 w-8 text-black" />
-              </motion.div>
-              <h1 className="text-2xl font-light ml-3 text-black">
-                Upload Avatar
-              </h1>
-            </div>
-
-            <AnimatePresence mode="wait">
-              {blob ? (
-                <motion.div
-                  key="success"
-                  initial={{ opacity: 0 }}
-                  animate={{ opacity: 1 }}
-                  exit={{ opacity: 0 }}
-                  className="text-center"
-                >
-                  <div className="flex flex-col items-center">
-                    <motion.div
-                      initial={{ scale: 0 }}
-                      animate={{ scale: 1 }}
-                      transition={{
-                        type: "spring",
-                        stiffness: 260,
-                        damping: 20,
-                      }}
-                      className="mb-6"
-                    >
-                      <CheckCircle className="h-12 w-12 text-black" />
-                    </motion.div>
-                    <h2 className="text-lg font-medium mb-4 text-black">
-                      Upload Complete
-                    </h2>
-
-                    <div className="mb-6 overflow-hidden rounded-full border border-gray-300">
-                      <img
-                        src={blob.url}
-                        alt="Uploaded avatar"
-                        className="h-24 w-24 object-cover"
-                      />
-                    </div>
-
-                    <div className="mb-6 p-3 bg-gray-50 border border-gray-200 rounded text-xs break-all">
-                      <p className="font-mono text-gray-600">
-                        {blob.url}
-                      </p>
-                    </div>
-
-                    <div className="flex space-x-3">
-                      <a
-                        href={blob.url}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="px-4 py-2 bg-black text-white text-sm rounded hover:bg-gray-800 transition-colors"
-                      >
-                        View
-                      </a>
-                      <button
-                        onClick={handleReset}
-                        className="px-4 py-2 border border-gray-300 text-black text-sm rounded hover:bg-gray-50 transition-colors"
-                      >
-                        Upload Another
-                      </button>
-                    </div>
-                  </div>
-                </motion.div>
-              ) : (
-                <motion.form
-                  key="upload-form"
-                  initial={{ opacity: 0 }}
-                  animate={{ opacity: 1 }}
-                  exit={{ opacity: 0 }}
-                  onSubmit={handleSubmit}
-                  className="space-y-6"
-                >
-                  <div
-                    className={`relative border-2 border-dashed rounded-lg p-8 transition-colors ${
-                      dragActive
-                        ? "border-black bg-gray-50"
-                        : "border-gray-300"
-                    }`}
-                    onDragEnter={handleDrag}
-                    onDragLeave={handleDrag}
-                    onDragOver={handleDrag}
-                    onDrop={handleDrop}
-                  >
-                    <input
-                      name="file"
-                      ref={inputFileRef}
-                      type="file"
-                      required
-                      accept="image/*"
-                      onChange={handleInputChange}
-                      className="absolute inset-0 w-full h-full opacity-0 cursor-pointer z-10"
-                    />
-
-                    <div className="flex flex-col items-center justify-center text-center">
-                      <AnimatePresence mode="wait">
-                        {preview ? (
-                          <motion.div
-                            key="preview"
-                            initial={{ opacity: 0, scale: 0.9 }}
-                            animate={{ opacity: 1, scale: 1 }}
-                            exit={{ opacity: 0, scale: 0.9 }}
-                            className="relative mb-4 group"
-                          >
-                            <div className="h-24 w-24 rounded-full overflow-hidden border border-gray-300 dark:border-gray-700">
-                              <img
-                                src={preview}
-                                alt="Avatar preview"
-                                className="h-full w-full object-cover"
-                              />
-                            </div>
-                            <button
-                              type="button"
-                              onClick={handleReset}
-                              className="absolute -top-1 -right-1 bg-black dark:bg-white text-white dark:text-black rounded-full p-1 opacity-0 group-hover:opacity-100 transition-opacity"
-                            >
-                              <X className="h-3 w-3" />
-                            </button>
-                          </motion.div>
-                        ) : (
-                          <motion.div
-                            key="upload-icon"
-                            initial={{ opacity: 0 }}
-                            animate={{ opacity: 1 }}
-                            exit={{ opacity: 0 }}
-                          >
-                            <Upload className="h-12 w-12 text-gray-400 dark:text-gray-600 mb-4" />
-                          </motion.div>
-                        )}
-                      </AnimatePresence>
-
-                      <h3 className="text-base font-medium mb-1 text-black dark:text-white">
-                        {preview ? "Ready to upload" : "Drop image here"}
-                      </h3>
-                      <p className="text-sm text-gray-500 dark:text-gray-400 mb-2">
-                        {preview
-                          ? "Click upload to continue"
-                          : "or click to browse files"}
-                      </p>
-                      <p className="text-xs text-gray-400 dark:text-gray-500">
-                        PNG, JPG, WEBP • Max 5MB
-                      </p>
-                    </div>
-                  </div>
-
-                  <AnimatePresence>
-                    {error && (
-                      <motion.div
-                        initial={{ opacity: 0, y: -10 }}
-                        animate={{ opacity: 1, y: 0 }}
-                        exit={{ opacity: 0 }}
-                        className="flex items-center p-3 rounded border border-gray-300 dark:border-gray-700 bg-gray-50 dark:bg-gray-900 text-black dark:text-white"
-                      >
-                        <AlertCircle className="h-4 w-4 mr-2 flex-shrink-0" />
-                        <p className="text-sm">{error}</p>
-                      </motion.div>
-                    )}
-                  </AnimatePresence>
-
-                  <motion.button
-                    whileHover={{ scale: 1.01 }}
-                    whileTap={{ scale: 0.99 }}
-                    type="submit"
-                    disabled={loading || !preview}
-                    className={`w-full py-3 px-4 rounded font-medium transition-all
-                      ${
-                        preview
-                          ? "bg-black dark:bg-white text-white dark:text-black hover:bg-gray-800 dark:hover:bg-gray-200"
-                          : "bg-gray-100 dark:bg-gray-900 border border-gray-300 dark:border-gray-700 text-gray-400 dark:text-gray-600 cursor-not-allowed"
-                      }
-                      ${loading ? "opacity-70 cursor-wait" : ""}
-                    `}
-                  >
-                    {loading ? (
-                      <div className="flex justify-center items-center">
-                        <div className="h-4 w-4 border-2 border-t-transparent border-current rounded-full animate-spin mr-2"></div>
-                        Uploading
-                      </div>
-                    ) : (
-                      "Upload Avatar"
-                    )}
-                  </motion.button>
-                </motion.form>
-              )}
-            </AnimatePresence>
-          </div>
-        </motion.div>
-      </div>
-    </div>
-  );
+export default function ImageUploadPage() {
+  return <UploadClientPage />
 }
