+++ app/layout.tsx
@@ -1,25 +1,26 @@
-import type React from "react"
-import type { Metadata } from "next"
-import { Inter } from "next/font/google"
-import { Analytics } from "@vercel/analytics/react"
-import { SpeedInsights } from "@vercel/speed-insights/next"
-import { Suspense } from "react"
-import GoogleAnalytics from "@/_analytics/provider"
+import type React from "react";
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import { Analytics } from "@vercel/analytics/react";
+import { SpeedInsights } from "@vercel/speed-insights/next";
+import { Suspense } from "react";
+import GoogleAnalytics from "@/_analytics/provider";
 
-import { ThemeProvider } from "@/components/theme-provider"
-import { Header } from "@/components/header"
-import { Footer } from "@/components/footer"
-import { siteConfig } from "@/config/site"
-import { cn } from "@/lib/utils"
-import { ThemeEffect } from "@/components/theme-effect"
+import { ThemeProvider } from "@/components/theme-provider";
+import { Header } from "@/components/header";
+import { Footer } from "@/components/footer";
+import { siteConfig } from "@/config/site";
+import { cn } from "@/lib/utils";
+import { ThemeEffect } from "@/components/theme-effect";
+import { ReactScan } from "@/components/react-scan";
 
-import "./globals.css"
+import "./globals.css";
 
 const inter = Inter({
   subsets: ["latin"],
   display: "swap",
   variable: "--font-inter",
-})
+});
 
 export const metadata: Metadata = {
   metadataBase: new URL(siteConfig.url),
@@ -65,25 +66,38 @@ export const metadata: Metadata = {
     apple: "/apple-touch-icon.png",
   },
   manifest: "/site.webmanifest",
-    generator: 'v0.dev'
-}
+  generator: "Portfolio",
+};
 
 export default function RootLayout({
   children,
 }: Readonly<{
-  children: React.ReactNode
+  children: React.ReactNode;
 }>) {
   return (
     <html lang="en" suppressHydrationWarning>
       <head />
-      <body className={cn("min-h-screen bg-background font-sans antialiased theme-transition", inter.variable)}>
-        <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange={false}>
+      <ReactScan />
+      <body
+        className={cn(
+          "min-h-screen bg-background font-sans antialiased theme-transition",
+          inter.variable
+        )}
+      >
+        <ThemeProvider
+          attribute="class"
+          defaultTheme="system"
+          enableSystem
+          disableTransitionOnChange={false}
+        >
           <ThemeEffect />
           <div className="relative flex min-h-screen flex-col">
             <Header />
             <main className="flex-1 flex justify-center">
               <Suspense>
-                <div className="w-[40%] min-w-[320px] px-4 py-12">{children}</div>
+                <div className="w-[40%] min-w-[320px] px-4 py-12">
+                  {children}
+                </div>
               </Suspense>
             </main>
             <Footer />
@@ -94,5 +108,5 @@ export default function RootLayout({
         <SpeedInsights />
       </body>
     </html>
-  )
+  );
 }
