+++ components/react-scan.tsx
@@ -1,0 +1,13 @@
+"use client";
+import { scan } from "react-scan";
+import { useEffect } from "react";
+
+export function ReactScan() {
+  useEffect(() => {
+    scan({
+      enabled: process.env.NODE_ENV === "development",
+    });
+  }, []);
+
+  return null;
+}
