+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-v0-project",
+  "name": "Portfolio",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -64,6 +64,7 @@
     "react-dom": "^19",
     "react-hook-form": "^7.54.1",
     "react-resizable-panels": "^2.1.7",
+    "react-scan": "^0.3.4",
     "recharts": "2.15.0",
     "sonner": "latest",
     "tailwind-merge": "^2.5.5",
